Move (Not)ExistsSubquery to the right location
