+++ ibis/backends/base/sql/alchemy/registry.py
@@ -10,7 +10,6 @@ import ibis.expr.datatypes as dt
 import ibis.expr.operations as ops
 import ibis.expr.types as ir
 import ibis.expr.window as W
-from ibis.backends.base.sql.compiler import ExistsSubquery, NotExistsSubquery
 
 from .database import AlchemyTable
 from .geospatial import geospatial_supported
@@ -127,7 +126,7 @@ def _exists_subquery(t, expr):
     sub_ctx = ctx.subcontext()
     clause = to_sqlalchemy(filtered, sub_ctx, exists=True)
 
-    if isinstance(op, NotExistsSubquery):
+    if isinstance(op, ops.NotExistsSubquery):
         clause = sa.not_(clause)
 
     return clause
@@ -425,8 +424,8 @@ sqlalchemy_operation_registry = {
     ops.SearchedCase: _searched_case,
     ops.TableColumn: _table_column,
     ops.TableArrayView: _table_array_view,
-    ExistsSubquery: _exists_subquery,
-    NotExistsSubquery: _exists_subquery,
+    ops.ExistsSubquery: _exists_subquery,
+    ops.NotExistsSubquery: _exists_subquery,
     # miscellaneous varargs
     ops.Least: varargs(sa.func.least),
     ops.Greatest: varargs(sa.func.greatest),
