+++ ibis/backends/base/sql/compiler/query_builder.py
@@ -4,13 +4,91 @@ import toolz
 
 import ibis.common.exceptions as com
 import ibis.expr.analysis as L
+import ibis.expr.datatypes as dt
 import ibis.expr.operations as ops
 import ibis.expr.types as ir
 import ibis.util as util
 
-from . import transforms
 from .base import DML, QueryAST, SetOp
 from .extract_subqueries import ExtractSubqueries
+
+
+class _AnyToExistsTransform:
+
+    """
+    Some code duplication with the correlated ref check; should investigate
+    better code reuse.
+    """
+
+    def __init__(self, context, expr, parent_table):
+        self.context = context
+        self.expr = expr
+        self.parent_table = parent_table
+        self.query_roots = frozenset(self.parent_table._root_tables())
+
+    def get_result(self):
+        self.foreign_table = None
+        self.predicates = []
+
+        self._visit(self.expr)
+
+        if type(self.expr.op()) == ops.Any:
+            op = ops.ExistsSubquery(self.foreign_table, self.predicates)
+        else:
+            op = ops.NotExistsSubquery(self.foreign_table, self.predicates)
+
+        expr_type = dt.boolean.column_type()
+        return expr_type(op)
+
+    def _visit(self, expr):
+        node = expr.op()
+
+        for arg in node.flat_args():
+            if isinstance(arg, ir.TableExpr):
+                self._visit_table(arg)
+            elif isinstance(arg, ir.BooleanColumn):
+                for sub_expr in L.flatten_predicate(arg):
+                    self.predicates.append(sub_expr)
+                    self._visit(sub_expr)
+            elif isinstance(arg, ir.Expr):
+                self._visit(arg)
+            else:
+                continue
+
+    def _find_blocking_table(self, expr):
+        node = expr.op()
+
+        if node.blocks():
+            return expr
+
+        for arg in node.flat_args():
+            if isinstance(arg, ir.Expr):
+                result = self._find_blocking_table(arg)
+                if result is not None:
+                    return result
+
+    def _visit_table(self, expr):
+        node = expr.op()
+
+        if isinstance(expr, ir.TableExpr):
+            base_table = self._find_blocking_table(expr)
+            if base_table is not None:
+                base_node = base_table.op()
+                if self._is_root(base_node):
+                    pass
+                else:
+                    # Foreign ref
+                    self.foreign_table = expr
+        else:
+            if not node.blocks():
+                for arg in node.flat_args():
+                    if isinstance(arg, ir.Expr):
+                        self._visit(arg)
+
+    def _is_root(self, what):
+        if isinstance(what, ir.Expr):
+            what = what.op()
+        return what in self.query_roots
 
 
 class _CorrelatedRefCheck:
@@ -65,11 +143,7 @@ class _CorrelatedRefCheck:
         # XXX
         if isinstance(
             node,
-            (
-                ops.TableArrayView,
-                transforms.ExistsSubquery,
-                transforms.NotExistsSubquery,
-            ),
+            (ops.TableArrayView, ops.ExistsSubquery, ops.NotExistsSubquery,),
         ):
             return True
 
@@ -989,9 +1063,7 @@ class SelectBuilder:
     def _visit_filter_Any(self, expr):
         # Rewrite semi/anti-join predicates in way that can hook into SQL
         # translation step
-        transform = transforms.AnyToExistsTransform(
-            self.context, expr, self.table_set
-        )
+        transform = _AnyToExistsTransform(self.context, expr, self.table_set)
         return transform.get_result()
 
     _visit_filter_NotAny = _visit_filter_Any
